+++ TabletopTweaks-Base/NewContent/Units/CarnivorousCrystal.cs
@@ -2,7 +2,6 @@
 using Kingmaker.RuleSystem;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
 using Kingmaker.UnitLogic.Abilities.Components;
-using Kingmaker.UnitLogic.Buffs.Blueprints;
 using Kingmaker.UnitLogic.Mechanics;
 using Kingmaker.UnitLogic.Mechanics.Actions;
 using TabletopTweaks.Core.Utilities;
@@ -11,7 +10,7 @@ using static TabletopTweaks.Base.Main;
 namespace TabletopTweaks.Base.NewContent.Units {
     static class CarnivorousCrystal {
         public static void AddCarnivorousCrystalApplyImmunity() {
-            var CarnivorousCrystalBuffSubsonicHumImmunity = BlueprintTools.GetBlueprint<BlueprintBuff>("88e345f3233c8024e9d191a807c40223");
+            var CarnivorousCrystalBuffSubsonicHumImmunity = BlueprintTools.GetBlueprintReference<BlueprintBuffReference>("88e345f3233c8024e9d191a807c40223");
 
             var CarnivorousCrystalApplyImmunityAbility = Helpers.CreateBlueprint<BlueprintAbility>(TTTContext, "CarnivorousCrystalApplyImmunityAbility", bp => {
                 bp.SetName(TTTContext, "Apply Subsonic Hum Immunity");
@@ -30,7 +29,7 @@ namespace TabletopTweaks.Base.NewContent.Units {
                 bp.AddComponent<AbilityEffectRunAction>(c => {
                     c.Actions = Helpers.CreateActionList(
                         new ContextActionApplyBuff() {
-                            m_Buff = CarnivorousCrystalBuffSubsonicHumImmunity.ToReference<BlueprintBuffReference>(),
+                            m_Buff = CarnivorousCrystalBuffSubsonicHumImmunity,
                             DurationValue = new ContextDurationValue() {
                                 m_IsExtendable = true,
                                 Rate = DurationRate.Hours,